backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
groq==0.11.0
python-dotenv==1.0.0
